Add build script for Archer2
We can mostly compile AxiSEM-3D using libraries provided on the
system. However, the latest boost is (one minor version) too old
so that needs to be downloaded. There is also an interesting
issue with the naming scheme of metis, which is avoided by making
a local soft link to the library with a more common name
(otherwise CMake complains). Finally, it is important to use the
gcc compiler as the Cray C++ compiler throws an ICE which we have
not investigated.